% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Columns.R
\name{Columns}
\alias{Columns}
\title{Retrieve the names of available columns in a table.}
\usage{
Columns(connection, schema, table)
}
\arguments{
\item{connection}{A DBI Connection object, as returned by \code{odbc::dbConnect()}
and \code{Connect()}.}

\item{schema}{A character vector specifying the name of a schema. Valid
schema names can be identified using \code{Schemas()}.}

\item{table}{A character vector specifying the name of a table. Valid
table names can be identified using \code{Tables()}.}
}
\value{
A character vector containing the names of all visible columns in the
specified table in the correct order.
}
\description{
Fast return of the names of all visible columns in the specified
table in the correct order.
}
\examples{
 \dontrun{
 conn <- Connect(name = "data_source_name")

 Columns(conn, "my_schema", "my_table_name")
 }

}
