% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disconnect.R
\name{Disconnect}
\alias{Disconnect}
\title{Close an active connection to an ODBC DBMS.}
\usage{
Disconnect(connection)
}
\arguments{
\item{connection}{A DBI Connection object, as returned by \code{odbc::dbConnect()}
or \code{Connect()}.}
}
\description{
A wrapper for \code{odbc::dbDisconnect()}; closes a DB connection, as
created by \code{odbc::dbConnect()} or \code{Connect()}.
}
\examples{
 \dontrun{
 conn <- Connect("data_source_name")

 Disconnect(conn)
 }

}
