% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fields.R
\name{Fields}
\alias{Fields}
\title{Retrieve the names of available fields in a table.}
\usage{
Fields(connection, schema, table)
}
\arguments{
\item{connection}{A DBI Connection object, as returned by \code{odbc::dbConnect()}
and \code{Connect()}.}

\item{schema}{A character vector specifying the name of a schema. Valid
schema names can be identified by running \code{Schemas()} and noting rows with
"type" "schema".}

\item{table}{A character vector specifying the name of a table. Valid
table names can be identified by running \code{Tables()}.}
}
\value{
A character vector containing the names of all visible fields in the
specified table in the correct order.
}
\description{
Fast return of the names of all visible fields in the specified
table in the correct order.
}
\examples{
 \dontrun{
 conn <- Connect("data_source_name")

 Fields(conn, "my_schema", "my_table_name")
 }

}
