% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{Query}
\alias{Query}
\title{Retrieve the result of a SQL query as a data frame.}
\usage{
Query(connection, sql_query)
}
\arguments{
\item{connection}{A DBIConnection object, as returned by \code{odbc::dbConnect()}
and \code{Connect()}.}

\item{sql_query}{A character string containing SQL.}
}
\value{
A character vector of all visible fields in the specified
table in the correct order.
}
\description{
A Wrapper for \code{odbc::dbGetQuery()}. Returns the result of a
query as a data frame. The function is suitable for SELECT queries only.
}
\examples{
 \dontrun{
 conn <- Connect("data_source_name")

 Query(conn, "SELECT * FROM xx.yyy")
 }

}
