% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{Tables}
\alias{Tables}
\title{Retrieve the names of available tables.}
\usage{
Tables(connection, schema = NULL, contains = NULL)
}
\arguments{
\item{connection}{A DBIConnection object, as returned by \code{odbc::dbConnect()}
and \code{Connect()}.}

\item{schema}{A character string identifying a 'schema' for which the user
wants to return the names of tables within that schema. Valid schema names
can be identified by running \code{Schemas()} and noting rows with "type"
"schema".}

\item{contains}{A character string to be matched in the names of tables
available to the user, only schema with names containing the specified
string will be returned. Case is ignored.}
}
\value{
A character vector of the names of tables accessible through the
specified connection and/or schema.
}
\description{
A wrapper for \code{odbc::dbListTables()} which returns a character
vector containing the names of tables in alphabetical order. The  function
default returns the names of all visible tables within a connection, though
users can specify a schema to limit the return to tables within that schema.

Users can use the "contains" argument to limit the table names returned to
those with names containing a specified string.
}
\examples{
 \dontrun{
 conn <- Connect()

 Tables(conn)
 }

}
