% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{Connect}
\alias{Connect}
\title{Connect to an ODBC DBMS with secure authentication input prompts.}
\usage{
Connect(dsn = NULL, driver = NULL, host = NULL, port = NULL, trusted = FALSE)
}
\arguments{
\item{dsn}{Character string specifying the data source name (used
interchangeably to specify database name when ODBC configuration files are
not in place for the desired connection). The names of valid data sources
available to the user can be retrieved using the \code{ListDataSources()}
function.}

\item{driver}{Character string specifying the name of a driver to be used.}

\item{host}{Character string specifying the host name.}

\item{port}{Character string specifying  the port number.}

\item{trusted}{Logical value, use \code{TRUE} when the connection does not require
a username or password. The default, \code{FALSE}, will prompt the user for
credentials.}
}
\description{
Wrapper for \code{odbc::dbconnect()}, designed to make connecting to
an ODBC DBMS as simple as possible. The function is suitable for use in an
interactive RStudio session where the user will receive prompts to enter
credentials if required.

Credentials are handled securely using the {rstudioapi}. For users looking
to retrieve credentials from the system credential store using the {keyring}
package, please use \code{KeyringConnect()}.

Where ODBC configuration files are in place, users need only supply the
{dsn} argument to connect to a database.
}
\examples{
 \dontrun{
 conn <- Connect("data_source_name")
 }

}
