% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSources.R
\name{DataSources}
\alias{DataSources}
\title{Retrieve the names of data sources available to the user.}
\usage{
DataSources(contains = NULL)
}
\arguments{
\item{contains}{A character string to be matched in the names of data sources
available to the user, only data sources with names containing the specified
string will be returned. Case is ignored.}
}
\description{
A wrapper for \code{odbc::odbcListDataSources()} which returns a
character vector of the names of data sources available to the user. Users
can choose to return only data sources with names containing a specific
pattern by using the "contains" argument.

The data source names (dsn) returned are suitable for use in the "dsn"
argument of other functions within the package (e.g. \code{Connect()}).
}
\examples{
 \dontrun{
 DataSources()

 DataSources(contains = "patient")
 }

}
