% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schemas.R
\name{Schemas}
\alias{Schemas}
\title{Retrieve the names and types of all visible objects within a connection or
catalog.}
\usage{
Schemas(connection, catalog = NULL, contains = NULL)
}
\arguments{
\item{connection}{A DBI Connection object, as returned by \code{odbc::dbConnect()}
and \code{Connect()}.}

\item{catalog}{A character string identifying a 'catalog' object for which
the user wants to return the details of objects within that catalog. Valid
catalog names can be identified by running \code{Schemas()} and noting rows
with "type" "catalog".}

\item{contains}{A character string to be matched in the names of data sources
available to the user, only data sources with names containing the specified
string will be returned. Case is ignored.}
}
\value{
A data.frame containing the name and type of all visible objects
within a specified connection and/or catalog.
}
\description{
The default use returns a data.frame containing the name and
type of all visible objects within a connection. Where the user wants to
return the details of objects within a catalog, the catalog argument can be
used.

The typical usage is to identify available schema. Users can use the
"contains" argument to limit the rows returned to those with names
containing a specified string.
}
\examples{
 \dontrun{
 conn <- Connect()

 Schemas(conn)
 }

}
